requirements.txt
configparser==7.2.0
loguru==0.7.3
mysql-connector-python==9.5.0
